Convert ercole-agent in a CLI program
