+++ cmd/showConfig.go
@@ -1,0 +1,32 @@
+/*
+Copyright © 2022 NAME HERE <EMAIL ADDRESS>
+
+*/
+package cmd
+
+import (
+	"encoding/json"
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+// showConfigCmd represents the showConfig command
+var showConfigCmd = &cobra.Command{
+	Use:   "show-config",
+	Short: "Show the configuration",
+	Long:  `Show ercole-agent configuration`,
+	Run: func(cmd *cobra.Command, args []string) {
+		enc := json.NewEncoder(os.Stdout)
+		enc.SetIndent("", "    ")
+		if err := enc.Encode(configuration); err != nil {
+			fmt.Fprintf(os.Stderr, "%s\n", err)
+			os.Exit(1)
+		}
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(showConfigCmd)
+}
